Move code out of __init__.py (#30)
* adapted the dependency tree class to CLTKv1

* working on not reloading the full stanford pipeline on every process call

* cache stanford NLP objects in wrapper class

* fixed doctests for Stanford process

* automatically reformatted files

* implemented the pipeline pattern, and extraction of sentential structure of input text in the Stanford process

* implemented true pipelines and sentence extraction for stanfordNLP

* :

* moved code out of __init__.py
