+++ src/cltkv1/utils/pipelines.py
@@ -9,10 +9,10 @@ these dataclasses is to represent:
 from dataclasses import dataclass, field
 from typing import Callable, List, Type
 
-from cltkv1.languages import get_lang
+from cltkv1.languages.utils import get_lang
 from cltkv1.tokenizers import DefaultTokenizationProcess, LatinTokenizationProcess
 from cltkv1.utils.data_types import Language, Pipeline, Process
-from cltkv1.wrappers import StanfordNLPProcess
+from cltkv1.wrappers.stanford import StanfordNLPProcess
 
 
 @dataclass
@@ -28,7 +28,7 @@ class LatinPipeline(Pipeline):
     >>> a_pipeline.language.name
     'Latin'
     >>> a_pipeline.processes[0]
-    <class 'cltkv1.wrappers.StanfordNLPProcess'>
+    <class 'cltkv1.wrappers.stanford.StanfordNLPProcess'>
     """
 
     description: str = "Pipeline for the Latin language"
@@ -49,7 +49,7 @@ class GreekPipeline(Pipeline):
     >>> a_pipeline.language.name
     'Ancient Greek'
     >>> a_pipeline.processes[0]
-    <class 'cltkv1.wrappers.StanfordNLPProcess'>
+    <class 'cltkv1.wrappers.stanford.StanfordNLPProcess'>
     """
 
     description: str = "Pipeline for the Greek language"
@@ -70,7 +70,7 @@ class OCSPipeline(Pipeline):
     >>> a_pipeline.language.name
     'Church Slavic'
     >>> a_pipeline.processes[0]
-    <class 'cltkv1.wrappers.StanfordNLPProcess'>
+    <class 'cltkv1.wrappers.stanford.StanfordNLPProcess'>
     """
 
     description: str = "Pipeline for the Old Church Slavonic language"
@@ -91,7 +91,7 @@ class OldFrenchPipeline(Pipeline):
     >>> a_pipeline.language.name
     'Old French (842-ca. 1400)'
     >>> a_pipeline.processes[0]
-    <class 'cltkv1.wrappers.StanfordNLPProcess'>
+    <class 'cltkv1.wrappers.stanford.StanfordNLPProcess'>
     """
 
     description: str = "Pipeline for the Old French language"
@@ -112,7 +112,7 @@ class GothicPipeline(Pipeline):
     >>> a_pipeline.language.name
     'Gothic'
     >>> a_pipeline.processes[0]
-    <class 'cltkv1.wrappers.StanfordNLPProcess'>
+    <class 'cltkv1.wrappers.stanford.StanfordNLPProcess'>
     """
 
     description: str = "Pipeline for the Gothic language"
